A little more cleanup. EntityCompare looked to be the start of an API of sorts, but was unused. No debug messages contain color codes, useless to try to strip them.
